Save/Open should not be managed by SideBar + identify beta features
