+++ src/App.css
@@ -5,27 +5,35 @@ html, body, #root {
 }
 
 body {
-  background: rgba(0,0,0,.75);
-  color: #fff;
+    background: rgba(0,0,0,.75);
+    color: #fff;
 }
 
 #wrapper {
-  height: 100%;
-  position: relative;
-  width: 100%;
+    height: 100%;
+    position: relative;
+    width: 100%;
 }
 
 #sidebar {
-  width: 30%;
-  min-height: 100%;
-  padding: 1%;
+    width: 30%;
+    min-height: 100%;
+    padding: 0;
+}
+
+.navbar {
+    margin-top: 0;
+    margin-bottom: 20px;
+    background: darkgrey;
+    color: #fff;
+    padding-left: 5%;
 }
 
 #app-maparea {
-  position: relative;
-  width: 70%;
-  height: inherit;
-  padding: 0;
+    position: relative;
+    width: 70%;
+    height: inherit;
+    padding: 0;
 }
 
 .leaflet-pane {
@@ -51,6 +59,10 @@ body {
     height: 100%;
     top: 0;
     padding: 0;
+}
+
+.accordion {
+    margin: 0 3%;
 }
 
 .card {
@@ -90,4 +102,18 @@ p.help {
 h4 {
     margin-top: 1.1em;
     color: #aaa;
+}
+
+.dropdown-item {
+    position: relative;
+}
+
+.beta::after {
+    content: "(beta)";
+    color: goldenrod;
+    font-size: 12px;
+    font-weight: bold;
+    padding: 4px;
+    position: absolute;
+    bottom: 3px;
 }
