+++ src/components/layers/layer.js
@@ -581,6 +581,7 @@ class Layer extends Component {
 									onChange={this.handleLayerTypeChange}
 									name="layerTypeSpatial"
 									disabled={ !this.state.hasSpatialPlugin }
+									className="beta"
 								/>
 							</Form.Group>
 
@@ -619,6 +620,7 @@ class Layer extends Component {
 									checked={this.state.rendering === RENDERING_HEATMAP}
 									onChange={this.handleRenderingChange}
 									name="mapRenderingHeatmap"
+									className="beta"
 								/>
 								<Form.Check
 									type="radio"
